Refactor settings panels into logical groups with standard size and tooltips.
- Created new UserControls: ApplicationSettings, DisplaySettings, SearchBehaviorSettings.
- Updated existing UserControls: LoggingSettings, CleanSqlSettings, DatabaseSettings, LauncherSettings.
- Standardized panel size to 450x400 with AutoScroll.
- Added descriptive labels and tooltips to all settings.
- Reorganized frmConfig navigation tree: General (Application, Display, Logging), Search (Behavior, Clean SQL), Database, Launcher.
- Removed obsolete settings files.
